tokenExtractor + userExtractor middlewares, delete succeeds only for user's own(added) blogs, broken test are fixed

+++ blog_list/controllers/blogs.js
@@ -2,6 +2,8 @@ const blogsRouter = require('express').Router()
 const { request, response } = require('../app')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -9,32 +11,48 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
   
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
 
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET) // The correctness of the token is verified + token is decoded (includes username & id)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id
+    title: title,
+    author: author,
+    url: url,
+    likes: likes || 0,
+    user: user.id
   })
   
-  if (blog.title !== undefined || blog.title !== "" || blog.url !== undefined || blog.url !== "" ) {
+  if (!(title && url)) {
+    response.status(400).json({ error: 'Title or url is missing' })
+  } else {
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
     response.status(201).json(savedBlog)
-  } else {
-    response.status(400).end()
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  
+  if (!blog) {
+    response.status(404).end()
+  }
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({error: 'Only the user who created the blog can delete it'})
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
